Adicionando o checkin apartir da plataforma parcialmente

+++ src/use-cases/dashboard/views/Staffs.vue
@@ -1,11 +1,13 @@
 <script setup>
-import { onMounted, computed, watch } from "vue";
+import { onMounted, computed, ref, watch } from "vue";
 import { useStore } from "vuex";
 import { useStaffs } from "@/repositories/staffs-repository";
 import { toast } from "vue3-toastify"
 import BtnSpinner from "../components/spinners/BtnSpinner.vue";
+import DynamicSearch from "../components/dynamics/DynamicSearch.vue";
 const { getStaffs, loading: loadingStaffs } = useStaffs()
 const { deleteStaff, loading: loadingDelete } = useStaffs()
+import Swal from 'sweetalert2';
 
 const store = useStore()
 
@@ -16,10 +18,18 @@ const _staff = computed(() => {
 const event = computed(() => {
     return store.getters.event
 })
+const user = computed(() => {
+    return store.getters.currentUser
+})
 
 const staffs = computed(() => {
     return store.getters.staffs
 })
+
+const page = ref(1);
+const limit = ref(10); // Valor inicial do limite
+const status = ref(null);
+const q = ref(null);
 
 const statusBgColor = (status) => {
     switch (status) {
@@ -68,20 +78,47 @@ const generateStatusLegend = (status) => {
     }
 }
 
-const _deleteStaff = async (staffId, memberId, index) => {
-    await deleteStaff({
-        eventId: event.value._id,
-        memberId: memberId,
-        staffId: staffId
-    }).then(() => {
-        toast("Colaborador excluido!", {
-            theme: "colored",
-            position: "top-right",
-            autoClose: 2500,
-            type: 'success'
-        })
-        store.dispatch("removeStaffFromStaffs", index)
-    })
+const _deleteStaff = async (staffId, member, index) => {
+
+    const result = await Swal.fire({
+        title: 'Remover Cargo?',
+        html: `<p>Tem certeza que deseja remover o cargo de <strong>${member.full_name}</strong> deste evento?</p>
+           <p class="text-muted">Esta ação não pode ser desfeita.</p>`,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Sim, remover cargo!',
+        cancelButtonText: 'Cancelar',
+        reverseButtons: true,
+        showLoaderOnConfirm: true,
+        preConfirm: async () => {
+            await deleteStaff({
+                eventId: event.value._id,
+                memberId: member._id,
+                staffId: staffId
+            }).then(() => {
+                store.dispatch("removeStaffFromStaffs", index)
+            }).catch((error) => {
+                Swal.showValidationMessage(`Falha ao remover cargo!`);
+            });
+        },
+        allowOutsideClick: () => !Swal.isLoading()
+    });
+
+    if (result.isConfirmed) {
+        // Sucesso na remoção
+        Swal.fire({
+            title: 'Cargo Removido!',
+            text: `O cargo de ${member.full_name} foi removido com sucesso.`,
+            icon: 'success',
+            timer: 2000,
+            showConfirmButton: false
+        });
+
+        return true; // Retorna true indicando que foi removido
+    }
+
 }
 
 const openModal = (name, data) => {
@@ -92,11 +129,20 @@ const openModal = (name, data) => {
     })
 }
 
+const performSearch = () => {
+    page.value = 1; // Reseta a página ao mudar o limite ou buscar
+    fetchStaffs();
+};
+
 const fetchStaffs = async () => {
     await getStaffs({
         event: event.value._id,
-        page: 1,
-        limit: 10
+        page: page.value,
+        limit: limit.value,
+        ...(status.value && {
+            status: status.value
+        }),
+        q: q.value
     }).then((res) => {
         const data = res.data.staffs;
         const metadata = res.data.metadata;
@@ -108,6 +154,20 @@ const fetchStaffs = async () => {
         })
     })
 }
+
+const handlePrevPage = () => {
+    if (staffs?.value?.metadata?.hasPrevPage) {
+        page.value = staffs?.value?.metadata?.page - 1;
+        fetchStaffs();
+    }
+};
+
+const handleNextPage = () => {
+    if (staffs.value?.metadata?.hasNextPage) {
+        page.value = staffs?.value?.metadata?.page + 1;
+        fetchStaffs();
+    }
+};
 
 onMounted(async () => {
     if (event?.value?._id) {
@@ -124,77 +184,131 @@ watch(event, (newEvent) => {
 
 <template>
     <div>
-        <div class="flex justify-end mb-2">
-            <button v-if="_staff.role == 'manager'" @click="openModal('staff-form', { type: 'create' })"
-                class="px-4 py-2 bg-brand-primary text-xs font-semibold text-white rounded-sm duration-200">
-                Adicionar colaborador
-            </button>
-        </div>
-        <div class="bg-white p-4 mb-5 relative overflow-x-auto inline-block w-full shadow-md rounded-md">
-            <table class="w-full text-sm text-left border border-gray-100">
-                <thead class="text-xs border-b border-gray-100 text-gray-500 uppercase bg-gray-50">
-                    <tr>
-                        <th class="px-4 py-3">
-                            Status do convite
-                        </th>
-                        <th class="px-4 py-3">
-                            Colaborador
-                        </th>
-                        <th class="px-4 py-3">
-                            Nivel de Acesso
-                        </th>
-                        <th class="px-4 py-3">
+        <div class="bg-white p-4 mb-5 relative inline-block w-full shadow-md rounded-md">
+            <div class="flex flex-col lg:flex-row items-center justify-between mb-4 gap-4">
+                <div class="lg:max-w-80 w-full">
+                    <DynamicSearch v-model:search="q" :loading="loadingStaffs" @update:search="performSearch" />
+                </div>
+                <div class="lg:w-auto w-full">
+                    <button v-if="_staff.role == 'manager'" @click="openModal('staff-form', { type: 'create' })"
+                        class="px-4 py-2 text-sm font-medium text-white bg-brand-primary border border-transparent rounded-md focus:outline-none w-full focus:ring-offset-2 transition-colors duration-200 disabled:opacity-50 disabled:cursor-not-allowed">
+                        Novo colaborador(a)
+                    </button>
+                </div>
+            </div>
+            <div class="overflow-x-auto">
+                <table class="w-full text-sm text-left border border-gray-100">
+                    <thead class="text-xs border-b border-gray-100 text-gray-500 uppercase bg-gray-50">
+                        <tr>
+                            <th class="px-4 py-3">
+                                Status do convite
+                            </th>
+                            <th class="px-4 py-3">
+                                Colaborador
+                            </th>
+                            <th class="px-4 py-3">
+                                Nivel de Acesso
+                            </th>
+                            <th class="px-4 py-3">
 
-                        </th>
-                    </tr>
-                </thead>
-                <tbody v-if="!loadingStaffs">
-                    <tr v-if="staffs.data.length" v-for="(staff, index) in staffs.data" :key="staff._id"
-                        class=" bg-white border-b border-gray-100">
+                            </th>
+                        </tr>
+                    </thead>
+                    <tbody v-if="!loadingStaffs">
+                        <tr v-if="staffs.data.length" v-for="(staff, index) in staffs.data" :key="staff._id"
+                            class=" bg-white border-b border-gray-100">
 
-                        <td class="px-4 py-3" :class="statusTxtColor(staff.invite.status)">
-                            <div class="flex items-center gap-2">
-                                <div class="w-[12px] h-[12px] rounded-full" :class="statusBgColor(staff.invite.status)">
+                            <td class="px-4 py-3" :class="statusTxtColor(staff.invite.status)">
+                                <div class="flex items-center gap-2">
+                                    <div class="w-[12px] h-[12px] rounded-full"
+                                        :class="statusBgColor(staff.invite.status)">
+                                    </div>
+                                    <p>{{ generateStatusLegend(staff.invite.status) }}</p>
                                 </div>
-                                <p>{{ generateStatusLegend(staff.invite.status) }}</p>
-                            </div>
 
-                        </td>
+                            </td>
 
-                        <td class="px-4 py-3 text-nowrap text-gray-500">
-                            <p>{{ staff.member.full_name }}</p>
-                        </td>
+                            <td class="px-4 py-3 text-nowrap text-gray-500">
+                                <p>{{ staff.member.full_name }}</p>
+                            </td>
 
-                        <td class="px-4 py-3 text-gray-500 text-nowrap">
-                            {{ generateRoleLegend(staff.role) }}
-                        </td>
+                            <td class="px-4 py-3 text-gray-500 text-nowrap">
+                                {{ generateRoleLegend(staff.role) }}
+                            </td>
 
-                        <td v-if="_staff.role == 'manager' && _staff.member != staff.member._id && !staff.isCreator"
-                            class="px-4 py-3 flex items-center gap-2 text-gray-500">
-                            <button @click="openModal('staff-form', { staff: { ...staff, index }, type: 'edit' })"
-                                v-if="staff.status == 'accepted'">Editar</button>
-                            <button v-if="staff.status == 'pending'">Reenviar</button>
-                            <button :disabled="loadingDelete"
-                                @click="_deleteStaff(staff._id, staff.member._id, index)">Excluir</button>
-                        </td>
+                            <td v-if="_staff.role == 'manager' && _staff.member != staff.member._id && !staff.member._id !== event?.created_by"
+                                class="px-4 py-3 flex items-center gap-2 text-gray-500">
+                                <button @click="openModal('staff-form', { staff: { ...staff, index }, type: 'edit' })"
+                                    v-if="staff.status == 'accepted'">Editar</button>
+                                <button v-if="staff.status == 'pending'">Reenviar</button>
+                                <button
+                                    class="border border-red-500 hover:bg-red-500 hover:text-white transition-colors text-red-500 py-2 px-3 text-xs font-bold rounded-lg disabled:opacity-50 disabled:cursor-not-allowed disabled:pointer-events-none"
+                                    :disabled="loadingDelete" @click="_deleteStaff(staff._id, staff.member, index)">
+                                    Remover
+                                </button>
+                            </td>
 
-                        <td class="px-4 py-3 flex items-center gap-2 text-gray-500"
-                            v-if="_staff.member == staff.member._id">
-                            <button>Desfazer</button>
-                        </td>
-                    </tr>
-                    <tr v-else>
-                        <td class="px-4 py-3 text-nowrap">Nenhum colaborador encotrado</td>
-                    </tr>
-                </tbody>
-                <tbody v-else>
-                    <tr>
-                        <td class="p-4 mx-auto">
-                            <BtnSpinner />
-                        </td>
-                    </tr>
-                </tbody>
-            </table>
+                            <td class="px-4 py-3 flex items-center gap-2 text-gray-500"
+                                v-if="_staff.member == staff.member._id && event?.created_by?._id !== user._id">
+                                <button
+                                :disabled="staff.member._id === user._id"
+                                    class="border border-red-500 hover:bg-red-500 hover:text-white transition-colors text-red-500 py-2 px-3 text-xs font-bold rounded-lg disabled:opacity-50 disabled:cursor-not-allowed disabled:pointer-events-none">Desfazer
+                                </button>
+                            </td>
+                        </tr>
+                        <tr v-else>
+                            <td class="px-4 py-3 text-nowrap text-gray-500">Nenhum colaborador encotrado</td>
+                        </tr>
+                    </tbody>
+                    <tbody v-else>
+                        <tr>
+                            <td class="p-4 mx-auto">
+                                <BtnSpinner />
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+            </div>
+
+            <div class="w-full">
+                <div class="flex flex-col sm:flex-row items-center justify-between py-4 px-4">
+                    <!-- Informação de resultados -->
+                    <div class="flex items-center gap-2 mb-4 sm:mb-0">
+                        <p class="text-sm flex-1 text-gray-600">
+                            Mostrando
+                            <span class="font-medium">{{ staffs?.metadata?.page || 1 }}</span>
+                            a
+                            <span class="font-medium">{{ staffs?.metadata?.limit || 10 }}</span>
+                            de
+                            <span class="font-medium">{{ staffs?.metadata?.total || 0 }}</span>
+                            resultados
+                        </p>
+
+                    </div>
+
+
+                    <!-- Controles de paginação -->
+                    <div class="flex items-center space-x-2">
+                        <button @click="handlePrevPage" :disabled="!staffs?.metadata?.hasPrevPage"
+                            class="px-4 py-2 text-sm font-medium text-gray-700 bg-white border border-gray-300 rounded-md hover:bg-gray-50 focus:outline-none focus:ring-offset-2 transition-colors duration-200 disabled:opacity-50 disabled:cursor-not-allowed"
+                            :class="{ 'opacity-50 cursor-not-allowed': !staffs?.metadata?.hasPrevPage }">
+                            Anterior
+                        </button>
+                        <select v-model="limit" @change="performSearch"
+                            class="block w-auto px-[20px] py-2 text-sm font-medium shrink-0 text-gray-700 bg-white border border-gray-300 rounded-md focus:outline-none">
+                            <option value="10">10</option>
+                            <option value="15">15</option>
+                            <option value="20">20</option>
+                            <option value="30">30</option>
+                        </select>
+                        <button @click="handleNextPage" :disabled="!staffs?.metadata?.hasNextPage"
+                            class="px-4 py-2 text-sm font-medium text-white bg-brand-primary border border-transparent rounded-md focus:outline-none focus:ring-offset-2 transition-colors duration-200 disabled:opacity-50 disabled:cursor-not-allowed"
+                            :class="{ 'opacity-50 cursor-not-allowed': !staffs?.metadata?.hasNextPage }">
+                            Próximo
+                        </button>
+                    </div>
+                </div>
+            </div>
         </div>
     </div>
 </template>
